server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   groq
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
cffi==1.16.0
    # via cryptography
click==8.1.7
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
cryptography==42.0.8
    # via -r requirements.in
distro==1.9.0
    # via groq
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via fastapi
exceptiongroup==1.2.1
    # via anyio
fastapi==0.111.0
    # via -r requirements.in
fastapi-cli==0.0.4
    # via fastapi
greenlet==3.0.3
    # via sqlalchemy
groq==0.9.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via
    #   fastapi
    #   groq
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.4
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
orjson==3.10.5
    # via fastapi
pycparser==2.22
    # via cffi
pydantic==2.7.4
    # via
    #   -r requirements.in
    #   fastapi
    #   groq
pydantic-core==2.18.4
    # via pydantic
pygments==2.18.0
    # via rich
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.1
    # via uvicorn
rich==13.7.1
    # via typer
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   groq
    #   httpx
sqlalchemy==2.0.30
    # via -r requirements.in
starlette==0.37.2
    # via fastapi
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   anyio
    #   fastapi
    #   groq
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   uvicorn
ujson==5.10.0
    # via fastapi
uvicorn[standard]==0.30.1
    # via fastapi
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn
